disable classname prefixing
* OrientGraphFactory.setLabelAsClassName() to disable the default prefixing of classnames with V_ or E_
* OrientGraph.labelToClassName() and OrientGraph.classNameToLabel() helpers
* OrientGraph.getEdgeIndexedKeys() and OrientGraph.getVertexIndexedKeys() verify the super class
* OrientGraph.createClass() throws if creating the same class with a different base class
* test suites modified
